+++ libs/eslint-plugin-internal/src/index.mjs
@@ -2,6 +2,7 @@ import exampleScreenContainsExampleRule from './example-screen-contains-example/
 import exampleScreenDefaultRule from './example-screen-default/index.mjs';
 import figmaConnectImportsPackageMatchRule from './figma-connect-imports-package-match/index.mjs';
 import figmaConnectImportsRequiredRule from './figma-connect-imports-required/index.mjs';
+import figmaConnectNoSemicolonRule from './figma-connect-no-semicolon/index.mjs';
 import noDeprecatedJsdocRule from './no-deprecated-jsdoc/index.mjs';
 import safelySpreadPropsRule from './safely-spread-props/index.mjs';
 
@@ -14,6 +15,7 @@ const plugin = {
     'no-deprecated-jsdoc': noDeprecatedJsdocRule,
     'figma-connect-imports-required': figmaConnectImportsRequiredRule,
     'figma-connect-imports-package-match': figmaConnectImportsPackageMatchRule,
+    'figma-connect-no-semicolon': figmaConnectNoSemicolonRule,
   },
   configs: {},
 };
@@ -49,6 +51,7 @@ Object.assign(plugin.configs, {
     rules: {
       'internal/figma-connect-imports-required': 'error',
       'internal/figma-connect-imports-package-match': 'error',
+      'internal/figma-connect-no-semicolon': 'error',
     },
   },
 });