+++ packages/mobile/src/dots/__figma__/DotCount.figma.tsx
@@ -7,7 +7,7 @@ figma.connect(
   DotCount,
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=155%3A11976',
   {
-    imports: ["import { DotCount } from '@coinbase/cds-mobile/dots/DotCount';"],
+    imports: ["import { DotCount } from '@coinbase/cds-mobile/dots/DotCount'"],
     props: {
       count: figma.enum('type', {
         'single digit': 1,