+++ packages/mobile/src/tabs/__figma__/SegmentedTabs.figma.tsx
@@ -8,7 +8,7 @@ figma.connect(
   SegmentedTabs,
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=20859-2979&m=dev',
   {
-    imports: ["import { SegmentedTabs } from '@coinbase/cds-mobile/tabs/SegmentedTabs';"],
+    imports: ["import { SegmentedTabs } from '@coinbase/cds-mobile/tabs/SegmentedTabs'"],
     variant: { tabs: '2 tabs' },
     props: {
       activeTab: figma.enum('active state', {
@@ -34,7 +34,7 @@ figma.connect(
   SegmentedTabs,
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=20859-2979&m=dev',
   {
-    imports: ["import { SegmentedTabs } from '@coinbase/cds-mobile/tabs/SegmentedTabs';"],
+    imports: ["import { SegmentedTabs } from '@coinbase/cds-mobile/tabs/SegmentedTabs'"],
     variant: { tabs: '3 tabs' },
     props: {
       activeTab: figma.enum('active state', {
@@ -62,7 +62,7 @@ figma.connect(
   SegmentedTabs,
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=20859-3073&m=dev',
   {
-    imports: ["import { SegmentedTab } from '@coinbase/cds-mobile/tabs/SegmentedTab';"],
+    imports: ["import { SegmentedTab } from '@coinbase/cds-mobile/tabs/SegmentedTab'"],
     props: {
       id: figma.string('title'),
       label: figma.string('title'),