+++ packages/web/src/controls/__figma__/NativeTextArea.figma.tsx
@@ -8,8 +8,8 @@ figma.connect(
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=14089%3A46502',
   {
     imports: [
-      "import { TextInput } from '@coinbase/cds-web/controls/TextInput';",
-      "import { NativeTextArea } from '@coinbase/cds-web/controls/NativeTextArea';",
+      "import { TextInput } from '@coinbase/cds-web/controls/TextInput'",
+      "import { NativeTextArea } from '@coinbase/cds-web/controls/NativeTextArea'",
     ],
     props: {
       label: figma.boolean('show label', {