+++ packages/web/src/navigation/__figma__/NavigationTitle.figma.tsx
@@ -7,7 +7,7 @@ figma.connect(
   NavigationTitle,
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=1221-19383',
   {
-    imports: ["import { NavigationTitle } from '@coinbase/cds-web/navigation/NavigationTitle';"],
+    imports: ["import { NavigationTitle } from '@coinbase/cds-web/navigation/NavigationTitle'"],
     props: {
       children: figma.string('page title'),
     },