+++ packages/web/src/overlays/tooltip/__figma__/Tooltip.figma.tsx
@@ -9,8 +9,8 @@ figma.connect(
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=715%3A14162',
   {
     imports: [
-      "import { Tooltip } from '@coinbase/cds-web/overlays';",
-      "import { Button } from '@coinbase/cds-web/buttons/Button';",
+      "import { Tooltip } from '@coinbase/cds-web/overlays'",
+      "import { Button } from '@coinbase/cds-web/buttons/Button'",
     ],
     variant: { type: 'body' },
     props: {
@@ -42,8 +42,8 @@ figma.connect(
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=715%3A14162',
   {
     imports: [
-      "import { Tooltip } from '@coinbase/cds-web/overlays';",
-      "import { Button } from '@coinbase/cds-web/buttons/Button';",
+      "import { Tooltip } from '@coinbase/cds-web/overlays'",
+      "import { Button } from '@coinbase/cds-web/buttons/Button'",
     ],
     variant: { type: 'title + body' },
     props: {